Moved files from util to common
