+++ gulp/src/common/cookie.js
