+++ gulp/src/common/spec.js
