+++ gulp/src/common/validateEmail.js
