+++ gulp/src/manageusers/User.jsx
@@ -5,8 +5,8 @@ import _ from 'lodash-compat';
 import Button from 'react-bootstrap/lib/Button';
 import {Link} from 'react-router';
 
-import {validateEmail} from 'util/validateEmail';
-import {getCsrf} from 'util/cookie';
+import {validateEmail} from 'common/validateEmail';
+import {getCsrf} from 'common/cookie';
 
 import RolesMultiselect from './RolesMultiselect';
 import HelpText from './HelpText';
